update sandbox contract
