+++ contracts/sandbox/src/lib.rs
@@ -8,10 +8,8 @@ use near_sdk::{
 
 use dcap_qvl::{verify, QuoteCollateralV3};
 
+mod chainsig;
 mod collateral;
-mod ecdsa;
-mod external;
-mod utils;
 
 #[near(serializers = [json, borsh])]
 #[derive(Clone)]
@@ -54,7 +52,7 @@ impl Contract {
 
     /// will throw on client if worker agent is not registered with a codehash in self.approved_codehashes
     pub fn require_approved_codehash(&mut self) {
-        let worker = self.get_worker(env::predecessor_account_id());
+        let worker = self.get_agent(env::predecessor_account_id());
         require!(self.approved_codehashes.contains(&worker.codehash));
     }
 
@@ -100,15 +98,20 @@ impl Contract {
         true
     }
 
-    pub fn get_signature(&mut self, payload: Vec<u8>, path: String) -> Promise {
+    pub fn request_signature(
+        &mut self,
+        path: String,
+        payload: String,
+        key_type: String,
+    ) -> Promise {
         self.require_approved_codehash();
 
-        ecdsa::get_sig(payload, path, 0)
+        chainsig::request_signature(path, payload, key_type)
     }
 
     // views
 
-    pub fn get_worker(&self, account_id: AccountId) -> Worker {
+    pub fn get_agent(&self, account_id: AccountId) -> Worker {
         self.worker_by_account_id
             .get(&account_id)
             .expect("no worker found")
