--- contracts/sandbox/src/utils.rs
@@ -1,4 +1,0 @@
-pub fn vec_to_fixed<T, const N: usize>(v: Vec<T>) -> [T; N] {
-    v.try_into()
-        .unwrap_or_else(|v: Vec<T>| panic!("Expected a Vec of length {} but it was {}", N, v.len()))
-}
